# packages/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ClubDirectoryItemOut": {
        "properties": {
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "logo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logo Url"
          },
          "my_membership_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MembershipStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "primary_color": {
            "title": "Primary Color",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "sport_types": {
            "items": {
              "$ref": "#/components/schemas/Sport"
            },
            "title": "Sport Types",
            "type": "array"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "city",
          "sport_types",
          "logo_url",
          "primary_color",
          "timezone"
        ],
        "title": "ClubDirectoryItemOut",
        "type": "object"
      },
      "ClubDirectoryOut": {
        "description": "Datos públicos de un club para el directorio de la app.",
        "properties": {
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "logo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logo Url"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "primary_color": {
            "title": "Primary Color",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "sport_types": {
            "items": {
              "$ref": "#/components/schemas/Sport"
            },
            "title": "Sport Types",
            "type": "array"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "city",
          "sport_types",
          "logo_url",
          "primary_color",
          "timezone"
        ],
        "title": "ClubDirectoryOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MembershipStatus": {
        "enum": [
          "INVITED",
          "PENDING",
          "APPROVED",
          "REJECTED",
          "INACTIVE"
        ],
        "title": "MembershipStatus",
        "type": "string"
      },
...
      "MyMembershipOut": {
        "properties": {
          "club": {
            "$ref": "#/components/schemas/ClubDirectoryOut"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "joined_on": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Joined On"
          },
          "member_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Member Number"
          },
          "plan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlanSummaryOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "requested_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested At"
          },
          "status": {
            "$ref": "#/components/schemas/MembershipStatus"
          }
        },
        "required": [
          "id",
          "status",
          "member_number",
          "joined_on",
          "requested_at",
          "club",
          "plan"
        ],
        "title": "MyMembershipOut",
        "type": "object"
      },
...
      "PlanSummaryOut": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "monthly_fee": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Monthly Fee",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "monthly_fee"
        ],
        "title": "PlanSummaryOut",
        "type": "object"
      },
...
      "Sport": {
        "enum": [
          "tennis",
          "padel",
          "football",
          "basketball",
          "hockey",
          "volleyball",
          "rugby",
          "other"
        ],
        "title": "Sport",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "ClubSystem API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/mobile/clubs": {
      "get": {
        "description": "Clubes activos (hasta 100) con el estado de mi membresía en cada uno.",
        "operationId": "clubs_directory_api_v1_mobile_clubs_get",
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ClubDirectoryItemOut"
                  },
                  "title": "Response Clubs Directory Api V1 Mobile Clubs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Clubs Directory",
        "tags": [
          "App: clubes y membresías"
        ]
      }
    },
...
    "/api/v1/mobile/clubs/{club_id}/membership-requests": {
      "post": {
        "operationId": "request_membership_api_v1_mobile_clubs__club_id__membership_requests_post",
        "parameters": [
          {
            "in": "path",
            "name": "club_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Club Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyMembershipOut"
                }
              }
            },
            "description": "Ya existía una solicitud"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyMembershipOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Request Membership",
        "tags": [
          "App: clubes y membresías"
        ]
      }
    },
...
    "/api/v1/mobile/memberships/{membership_id}/accept": {
      "post": {
        "operationId": "accept_invitation_api_v1_mobile_memberships__membership_id__accept_post",
        "parameters": [
          {
            "in": "path",
            "name": "membership_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Membership Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyMembershipOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Accept Invitation",
        "tags": [
          "App: clubes y membresías"
        ]
      }
    },
    "/api/v1/mobile/memberships/{membership_id}/decline": {
      "post": {
        "operationId": "decline_invitation_api_v1_mobile_memberships__membership_id__decline_post",
        "parameters": [
          {
            "in": "path",
            "name": "membership_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Membership Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Decline Invitation",
        "tags": [
          "App: clubes y membresías"
        ]
      }
    },
...
  }
}